+++ tests/data/forms/nested/form.a2.ts
@@ -1,0 +1,12 @@
+
+import { Form } from "../../../../src";
+
+const fields = [
+    'user',
+    'user.firstname',
+    'user.lastname',
+];
+
+export default new Form({ fields }, { name: "Nested-A2", options: {
+    fallbackValue: null,
+} });